+++ components/calendar/index.ts
@@ -1,1 +1,3 @@
 export { TrainingCalendar, TrainingCalendarSkeleton } from "./TrainingCalendar";
+export { DraggableRunBadge, RunBadgeOverlay } from "./DraggableRunBadge";
+export { DroppableCalendarCell } from "./DroppableCalendarCell";